+++ templates/preact-app/src/index.js
@@ -1,16 +1,15 @@
+if (process.env.NODE_ENV === 'development') {
+  // Enable use of Preact Devtools
+  require('preact/debug')
+}
+
 import './index.css'
 
 import {h, render} from 'preact'
 
-if (process.env.NODE_ENV === 'development') {
-  // Enable use of React Developer Tools
-  require('preact/devtools')
-}
-
-let root
 function init() {
   let App = require('./App').default
-  root = render(<App/>, document.querySelector('#app'), root)
+  render(<App/>, document.querySelector('#app'))
 }
 
 if (module.hot) {